app: move GET /api/config to app.routers.config
